+++ Pandora/Paradigm/Structure/Ability/Morphable.hs
@@ -33,6 +33,8 @@ data Morph a = Rotate a | Into a | Insert a | Push a | Pop a | Delete a | Find a
 
 data Occurrence a = All a | First a
 
+data Vertical a = Up a | Down a
+
 rotate :: forall f t . Morphable (Rotate f) t => t ~> Morphing (Rotate f) t
 rotate = morphing . TU . Tag @(Rotate f)